refractor
- add support for win32 uuid
- divide sources for easier maintenance.
