+++ src/ezgrpc2_grpc_status.h
@@ -1,0 +1,31 @@
+#ifndef EZGRPC2_GRPC_STATUS_H
+#define EZGRPC2_GRPC_STATUS_H
+
+typedef enum ezgrpc2_grpc_status_t ezgrpc2_grpc_status_t;
+/**
+ * Types of trailer status codes
+ */
+enum ezgrpc2_grpc_status_t {
+  /* https://github.com/grpc/grpc/tree/master/include/grpcpp/impl/codegen */
+  EZGRPC2_STATUS_OK = 0,
+  EZGRPC2_STATUS_CANCELLED = 1,
+  EZGRPC2_STATUS_UNKNOWN = 2,
+  EZGRPC2_STATUS_INVALID_ARGUMENT = 3,
+  EZGRPC2_STATUS_DEADLINE_EXCEEDED = 4,
+  EZGRPC2_STATUS_NOT_FOUND = 5,
+  EZGRPC2_STATUS_ALREADY_EXISTS = 6,
+  EZGRPC2_STATUS_PERMISSION_DENIED = 7,
+  EZGRPC2_STATUS_RESOURCE_EXHAUSTED = 8,
+  EZGRPC2_STATUS_FAILED_PRECONDITION = 9,
+  EZGRPC2_STATUS_OUT_OF_RANGE = 11,
+  EZGRPC2_STATUS_UNIMPLEMENTED = 12,
+  EZGRPC2_STATUS_INTERNAL = 13,
+  EZGRPC2_STATUS_UNAVAILABLE = 14,
+  EZGRPC2_STATUS_DATA_LOSS = 15,
+  EZGRPC2_STATUS_UNAUTHENTICATED = 16,
+  EZGRPC2_STATUS_NULL = -1
+};
+
+char *ezgrpc2_grpc_status_strstatus(ezgrpc2_grpc_status_t status);
+
+#endif
